pyproject.toml
[project]
name = "universal-memory-service"
version = "0.1.0"
description = "Self-hosted universal memory service for AI agents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "httpx>=0.28",
    "pyyaml>=6.0",
    "watchfiles>=1.0",
    "aiofiles>=24.0",
    "aiosqlite>=0.20",
    "llama-cpp-python>=0.3",
    "google-genai>=1.0",
    "mcp>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/universal_memory"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100
